separate sql tests
